requirements.txt
beautifulsoup4==4.8.0
captcha-solver
django==2.2.1
gunicorn==19.7.1
lxml==4.4.1
Pillow==6.1.0
PySocks==1.7.0
requests==2.22.0
selenium==3.141.0